Updated read me updated error handling
